Major documentation and codebase modernization
- Overhauled and reorganized all documentation for Nalix, adding new guides, deep-dive concept pages, and improved navigation.
- Unified and clarified core concepts, networking, extensibility, deployment, and tools; added new pages for server boilerplate, blueprint, middleware, session resumption, serialization, and more.
- Updated all code samples and project files to match Nalix v12.1.0 APIs and best practices.
- Standardized Vietnamese diacritics and encoding in source files; added scripts for encoding checks/fixes.
- Improved application startup/shutdown logic, error handling, and logging.
- Updated handler/DTO mapping to use `ushort` for SequenceId.
- Added custom fonts and improved CSS for better docs readability.
- Removed obsolete files and improved project structure recommendations.
- Result: modern, maintainable, and production-ready docs and codebase focused on developer experience, security, and clarity.

+++ src/AutoX.Gara.Api.Handlers/Common/HandlerExtensions.cs
@@ -17,7 +17,7 @@ public static class HandlerExtensions
             ControlType.ERROR, 
             reason, 
             ProtocolAdvice.DO_NOT_RETRY, 
-            (uint)context.Packet.SequenceId);
+            context.Packet.SequenceId);
             
         await context.Connection.TCP.SendAsync(directive).ConfigureAwait(false);
     }
@@ -31,7 +31,7 @@ public static class HandlerExtensions
             ControlType.NONE, 
             ProtocolReason.NONE, 
             ProtocolAdvice.NONE, 
-            (uint)context.Packet.SequenceId);
+            context.Packet.SequenceId);
             
         await context.Connection.TCP.SendAsync(directive).ConfigureAwait(false);
     }
